test: switch utils/stylesheets tests to react testing library
